+++ pyLinguist/utils.py
@@ -1,3 +1,7 @@
+from collections import Callable
+from threading import Thread
+
+
 LEVELS = {
     "NOTSET": 0,
     "DEBUG": 10,
@@ -7,6 +11,18 @@ LEVELS = {
     "ERROR": 40,
     "FATAL": 50,
     "CRITICAL": 50
+}
+
+ERROR_CODES = {
+    400: "Wrong parameter was specified",
+    401: "Invalid API key",
+    402: "Blocked API key",
+    403: "Exceeded the daily limit on the amount of requests",
+    404: "Exceeded the daily limit on the amount of translated text",
+    413: "Exceeded the maximum text size",
+    422: "The text cannot be translated",
+    501: "The specified translation direction is not supported",
+    503: "Server not available",
 }
 
 
@@ -20,3 +36,11 @@ def Logger(name: str, level: int=LEVELS["DEBUG"], **kwargs) -> ...:
         level=level
     )
     return logging.getLogger(name)
+
+
+def async_run(f: Callable) -> Callable:
+    '''Run functions in separate threads'''
+    def wrapper(*args, **kwargs) -> None:
+        thread = Thread(target=f, args=args, kwargs=kwargs)
+        thread.start()
+    return wrapper
